+++ src/crypto/validate.rs
@@ -1,0 +1,908 @@
+//! The code here needs to be refactored.
+//!
+//! There are direct references to the ring crate that need to be replaced
+//! with references to common.
+
+use crate::base::iana::{DigestAlg, Nsec3HashAlg, SecAlg};
+use crate::base::rdata::ComposeRecordData;
+use crate::base::wire::{Compose, Composer};
+use crate::base::{CanonicalOrd, Name, Record, RecordData, ToName};
+use crate::dep::octseq::builder::with_infallible;
+use crate::dep::octseq::{EmptyBuilder, OctetsBuilder, Truncate};
+use crate::rdata::nsec3::{Nsec3Salt, OwnerHash};
+use crate::rdata::{Dnskey, Nsec3param, Rrsig};
+
+use bytes::Bytes;
+
+use ring::digest::SHA1_FOR_LEGACY_USE_ONLY;
+use ring::{digest, signature};
+
+use std::vec::Vec;
+use std::{error, fmt};
+
+//------------ Dnskey --------------------------------------------------------
+
+/// Extensions for DNSKEY record type.
+pub trait DnskeyExt {
+    /// Calculates a digest from DNSKEY.
+    ///
+    /// See [RFC 4034, Section 5.1.4]:
+    ///
+    /// ```text
+    /// 5.1.4.  The Digest Field
+    ///   The digest is calculated by concatenating the canonical form of the
+    ///   fully qualified owner name of the DNSKEY RR with the DNSKEY RDATA,
+    ///   and then applying the digest algorithm.
+    ///
+    ///     digest = digest_algorithm( DNSKEY owner name | DNSKEY RDATA);
+    ///
+    ///      "|" denotes concatenation
+    ///
+    ///     DNSKEY RDATA = Flags | Protocol | Algorithm | Public Key.
+    /// ```
+    ///
+    /// [RFC 4034, Section 5.1.4]: https://tools.ietf.org/html/rfc4034#section-5.1.4
+    fn digest<N: ToName>(
+        &self,
+        name: &N,
+        algorithm: DigestAlg,
+    ) -> Result<digest::Digest, AlgorithmError>;
+}
+
+impl<Octets> DnskeyExt for Dnskey<Octets>
+where
+    Octets: AsRef<[u8]>,
+{
+    /// Calculates a digest from DNSKEY.
+    ///
+    /// See [RFC 4034, Section 5.1.4]:
+    ///
+    /// ```text
+    /// 5.1.4.  The Digest Field
+    ///   The digest is calculated by concatenating the canonical form of the
+    ///   fully qualified owner name of the DNSKEY RR with the DNSKEY RDATA,
+    ///   and then applying the digest algorithm.
+    ///
+    ///     digest = digest_algorithm( DNSKEY owner name | DNSKEY RDATA);
+    ///
+    ///      "|" denotes concatenation
+    ///
+    ///     DNSKEY RDATA = Flags | Protocol | Algorithm | Public Key.
+    /// ```
+    ///
+    /// [RFC 4034, Section 5.1.4]: https://tools.ietf.org/html/rfc4034#section-5.1.4
+    fn digest<N: ToName>(
+        &self,
+        name: &N,
+        algorithm: DigestAlg,
+    ) -> Result<digest::Digest, AlgorithmError> {
+        let mut buf: Vec<u8> = Vec::new();
+        with_infallible(|| {
+            name.compose_canonical(&mut buf)?;
+            self.compose_canonical_rdata(&mut buf)
+        });
+
+        let mut ctx = match algorithm {
+            DigestAlg::SHA1 => {
+                digest::Context::new(&digest::SHA1_FOR_LEGACY_USE_ONLY)
+            }
+            DigestAlg::SHA256 => digest::Context::new(&digest::SHA256),
+            DigestAlg::SHA384 => digest::Context::new(&digest::SHA384),
+            _ => {
+                return Err(AlgorithmError::Unsupported);
+            }
+        };
+
+        ctx.update(&buf);
+        Ok(ctx.finish())
+    }
+}
+
+// This needs to match the digests supported in digest.
+pub fn supported_digest(d: &DigestAlg) -> bool {
+    *d == DigestAlg::SHA1
+        || *d == DigestAlg::SHA256
+        || *d == DigestAlg::SHA384
+}
+
+//------------ Rrsig ---------------------------------------------------------
+
+/// Extensions for DNSKEY record type.
+pub trait RrsigExt {
+    /// Compose the signed data according to [RC4035, Section 5.3.2](https://tools.ietf.org/html/rfc4035#section-5.3.2).
+    ///
+    /// ```text
+    ///    Once the RRSIG RR has met the validity requirements described in
+    ///    Section 5.3.1, the validator has to reconstruct the original signed
+    ///    data.  The original signed data includes RRSIG RDATA (excluding the
+    ///    Signature field) and the canonical form of the RRset.  Aside from
+    ///    being ordered, the canonical form of the RRset might also differ from
+    ///    the received RRset due to DNS name compression, decremented TTLs, or
+    ///    wildcard expansion.
+    /// ```
+    fn signed_data<N: ToName, D, B: Composer>(
+        &self,
+        buf: &mut B,
+        records: &mut [impl AsRef<Record<N, D>>],
+    ) -> Result<(), B::AppendError>
+    where
+        D: RecordData + CanonicalOrd + ComposeRecordData + Sized;
+
+    /// Return if records are expanded for a wildcard according to the
+    /// information in this signature.
+    fn wildcard_closest_encloser<N, D>(
+        &self,
+        rr: &Record<N, D>,
+    ) -> Option<Name<Bytes>>
+    where
+        N: ToName;
+
+    /// Attempt to use the cryptographic signature to authenticate the signed data, and thus authenticate the RRSET.
+    /// The signed data is expected to be calculated as per [RFC4035, Section 5.3.2](https://tools.ietf.org/html/rfc4035#section-5.3.2).
+    ///
+    /// [RFC4035, Section 5.3.2](https://tools.ietf.org/html/rfc4035#section-5.3.2):
+    /// ```text
+    /// 5.3.3.  Checking the Signature
+    ///
+    ///    Once the resolver has validated the RRSIG RR as described in Section
+    ///    5.3.1 and reconstructed the original signed data as described in
+    ///    Section 5.3.2, the validator can attempt to use the cryptographic
+    ///    signature to authenticate the signed data, and thus (finally!)
+    ///    authenticate the RRset.
+    ///
+    ///    The Algorithm field in the RRSIG RR identifies the cryptographic
+    ///    algorithm used to generate the signature.  The signature itself is
+    ///    contained in the Signature field of the RRSIG RDATA, and the public
+    ///    key used to verify the signature is contained in the Public Key field
+    ///    of the matching DNSKEY RR(s) (found in Section 5.3.1).  [RFC4034]
+    ///    provides a list of algorithm types and provides pointers to the
+    ///    documents that define each algorithm's use.
+    /// ```
+    fn verify_signed_data(
+        &self,
+        dnskey: &Dnskey<impl AsRef<[u8]>>,
+        signed_data: &impl AsRef<[u8]>,
+    ) -> Result<(), AlgorithmError>;
+}
+
+impl<Octets: AsRef<[u8]>, TN: ToName> RrsigExt for Rrsig<Octets, TN> {
+    fn signed_data<N: ToName, D, B: Composer>(
+        &self,
+        buf: &mut B,
+        records: &mut [impl AsRef<Record<N, D>>],
+    ) -> Result<(), B::AppendError>
+    where
+        D: RecordData + CanonicalOrd + ComposeRecordData + Sized,
+    {
+        // signed_data = RRSIG_RDATA | RR(1) | RR(2)...  where
+        //    "|" denotes concatenation
+        // RRSIG_RDATA is the wire format of the RRSIG RDATA fields
+        //    with the Signature field excluded and the Signer's Name
+        //    in canonical form.
+        self.type_covered().compose(buf)?;
+        self.algorithm().compose(buf)?;
+        self.labels().compose(buf)?;
+        self.original_ttl().compose(buf)?;
+        self.expiration().compose(buf)?;
+        self.inception().compose(buf)?;
+        self.key_tag().compose(buf)?;
+        self.signer_name().compose_canonical(buf)?;
+
+        // The set of all RR(i) is sorted into canonical order.
+        // See https://tools.ietf.org/html/rfc4034#section-6.3
+        records.sort_by(|a, b| {
+            a.as_ref().data().canonical_cmp(b.as_ref().data())
+        });
+
+        // RR(i) = name | type | class | OrigTTL | RDATA length | RDATA
+        for rr in records.iter().map(|r| r.as_ref()) {
+            // Handle expanded wildcards as per [RFC4035, Section 5.3.2]
+            // (https://tools.ietf.org/html/rfc4035#section-5.3.2).
+            let rrsig_labels = usize::from(self.labels());
+            let fqdn = rr.owner();
+            // Subtract the root label from count as the algorithm doesn't
+            // accomodate that.
+            let fqdn_labels = fqdn.iter_labels().count() - 1;
+            if rrsig_labels < fqdn_labels {
+                // name = "*." | the rightmost rrsig_label labels of the fqdn
+                buf.append_slice(b"\x01*")?;
+                match fqdn
+                    .to_cow()
+                    .iter_suffixes()
+                    .nth(fqdn_labels - rrsig_labels)
+                {
+                    Some(name) => name.compose_canonical(buf)?,
+                    None => fqdn.compose_canonical(buf)?,
+                };
+            } else {
+                fqdn.compose_canonical(buf)?;
+            }
+
+            rr.rtype().compose(buf)?;
+            rr.class().compose(buf)?;
+            self.original_ttl().compose(buf)?;
+            rr.data().compose_canonical_len_rdata(buf)?;
+        }
+        Ok(())
+    }
+
+    fn wildcard_closest_encloser<N, D>(
+        &self,
+        rr: &Record<N, D>,
+    ) -> Option<Name<Bytes>>
+    where
+        N: ToName,
+    {
+        // Handle expanded wildcards as per [RFC4035, Section 5.3.2]
+        // (https://tools.ietf.org/html/rfc4035#section-5.3.2).
+        let rrsig_labels = usize::from(self.labels());
+        let fqdn = rr.owner();
+        // Subtract the root label from count as the algorithm doesn't
+        // accomodate that.
+        let fqdn_labels = fqdn.iter_labels().count() - 1;
+        if rrsig_labels < fqdn_labels {
+            // name = "*." | the rightmost rrsig_label labels of the fqdn
+            Some(
+                match fqdn
+                    .to_cow()
+                    .iter_suffixes()
+                    .nth(fqdn_labels - rrsig_labels)
+                {
+                    Some(name) => Name::from_octets(Bytes::copy_from_slice(
+                        name.as_octets(),
+                    ))
+                    .unwrap(),
+                    None => fqdn.to_bytes(),
+                },
+            )
+        } else {
+            None
+        }
+    }
+
+    fn verify_signed_data(
+        &self,
+        dnskey: &Dnskey<impl AsRef<[u8]>>,
+        signed_data: &impl AsRef<[u8]>,
+    ) -> Result<(), AlgorithmError> {
+        let signature = self.signature().as_ref();
+        let signed_data = signed_data.as_ref();
+
+        // Caller needs to ensure that the signature matches the key, but enforce the algorithm match
+        if self.algorithm() != dnskey.algorithm() {
+            return Err(AlgorithmError::InvalidData);
+        }
+
+        // Note: Canonicalize the algorithm, otherwise matching named variants against Int(_) is not going to work
+        let sec_alg = SecAlg::from_int(self.algorithm().to_int());
+        match sec_alg {
+            SecAlg::RSASHA1
+            | SecAlg::RSASHA1_NSEC3_SHA1
+            | SecAlg::RSASHA256
+            | SecAlg::RSASHA512 => {
+                let (algorithm, min_bytes) = match sec_alg {
+                    SecAlg::RSASHA1 | SecAlg::RSASHA1_NSEC3_SHA1 => (
+                        &signature::RSA_PKCS1_1024_8192_SHA1_FOR_LEGACY_USE_ONLY,
+                        1024 / 8,
+                    ),
+                    SecAlg::RSASHA256 => (
+                        &signature::RSA_PKCS1_1024_8192_SHA256_FOR_LEGACY_USE_ONLY,
+                        1024 / 8,
+                    ),
+                    SecAlg::RSASHA512 => (
+                        &signature::RSA_PKCS1_1024_8192_SHA512_FOR_LEGACY_USE_ONLY,
+                        1024 / 8,
+                    ),
+                    _ => unreachable!(),
+                };
+
+                // The key isn't available in either PEM or DER, so use the
+                // direct RSA verifier.
+                let (e, n) = rsa_exponent_modulus(dnskey, min_bytes)?;
+                let public_key =
+                    signature::RsaPublicKeyComponents { n: &n, e: &e };
+                public_key
+                    .verify(algorithm, signed_data, signature)
+                    .map_err(|_| AlgorithmError::BadSig)
+            }
+            SecAlg::ECDSAP256SHA256 | SecAlg::ECDSAP384SHA384 => {
+                let algorithm = match sec_alg {
+                    SecAlg::ECDSAP256SHA256 => {
+                        &signature::ECDSA_P256_SHA256_FIXED
+                    }
+                    SecAlg::ECDSAP384SHA384 => {
+                        &signature::ECDSA_P384_SHA384_FIXED
+                    }
+                    _ => unreachable!(),
+                };
+
+                // Add 0x4 identifier to the ECDSA pubkey as expected by ring.
+                let public_key = dnskey.public_key().as_ref();
+                let mut key = Vec::with_capacity(public_key.len() + 1);
+                key.push(0x4);
+                key.extend_from_slice(public_key);
+
+                signature::UnparsedPublicKey::new(algorithm, &key)
+                    .verify(signed_data, signature)
+                    .map_err(|_| AlgorithmError::BadSig)
+            }
+            SecAlg::ED25519 => {
+                let key = dnskey.public_key();
+                signature::UnparsedPublicKey::new(&signature::ED25519, &key)
+                    .verify(signed_data, signature)
+                    .map_err(|_| AlgorithmError::BadSig)
+            }
+            _ => Err(AlgorithmError::Unsupported),
+        }
+    }
+}
+
+// This needs to match the algorithms supported in signed_data.
+pub fn supported_algorithm(a: &SecAlg) -> bool {
+    *a == SecAlg::RSASHA1
+        || *a == SecAlg::RSASHA1_NSEC3_SHA1
+        || *a == SecAlg::RSASHA256
+        || *a == SecAlg::RSASHA512
+        || *a == SecAlg::ECDSAP256SHA256
+}
+
+/// Return the RSA exponent and modulus components from DNSKEY record data.
+fn rsa_exponent_modulus(
+    dnskey: &Dnskey<impl AsRef<[u8]>>,
+    min_len: usize,
+) -> Result<(&[u8], &[u8]), AlgorithmError> {
+    let public_key = dnskey.public_key().as_ref();
+    if public_key.len() <= 3 {
+        return Err(AlgorithmError::InvalidData);
+    }
+
+    let (pos, exp_len) = match public_key[0] {
+        0 => (
+            3,
+            (usize::from(public_key[1]) << 8) | usize::from(public_key[2]),
+        ),
+        len => (1, usize::from(len)),
+    };
+
+    // Check if there's enough space for exponent and modulus.
+    if public_key[pos..].len() < pos + exp_len {
+        return Err(AlgorithmError::InvalidData);
+    };
+
+    // Check for minimum supported key size
+    if public_key[pos..].len() < min_len {
+        return Err(AlgorithmError::Unsupported);
+    }
+
+    Ok(public_key[pos..].split_at(exp_len))
+}
+
+//============ Error Types ===================================================
+
+//------------ AlgorithmError ------------------------------------------------
+
+/// An algorithm error during verification.
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub enum AlgorithmError {
+    Unsupported,
+    BadSig,
+    InvalidData,
+}
+
+//--- Display, Error
+
+impl fmt::Display for AlgorithmError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        f.write_str(match self {
+            AlgorithmError::Unsupported => "unsupported algorithm",
+            AlgorithmError::BadSig => "bad signature",
+            AlgorithmError::InvalidData => "invalid data",
+        })
+    }
+}
+
+impl error::Error for AlgorithmError {}
+
+//------------ Nsec3HashError -------------------------------------------------
+
+/// An error when creating an NSEC3 hash.
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
+pub enum Nsec3HashError {
+    /// The requested algorithm for NSEC3 hashing is not supported.
+    UnsupportedAlgorithm,
+
+    /// Data could not be appended to a buffer.
+    ///
+    /// This could indicate an out of memory condition.
+    AppendError,
+
+    /// The hashing process produced an invalid owner hash.
+    ///
+    /// See: [OwnerHashError](crate::rdata::nsec3::OwnerHashError)
+    OwnerHashError,
+
+    /// The hashing process produced a hash that already exists.
+    CollisionDetected,
+
+    /// The hash provider did not provide a hash for the given owner name.
+    MissingHash,
+}
+
+//--- Display
+
+impl std::fmt::Display for Nsec3HashError {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        match self {
+            Nsec3HashError::UnsupportedAlgorithm => {
+                f.write_str("Unsupported algorithm")
+            }
+            Nsec3HashError::AppendError => {
+                f.write_str("Append error: out of memory?")
+            }
+            Nsec3HashError::OwnerHashError => {
+                f.write_str("Hashing produced an invalid owner hash")
+            }
+            Nsec3HashError::CollisionDetected => {
+                f.write_str("Hash collision detected")
+            }
+            Nsec3HashError::MissingHash => {
+                f.write_str("Missing hash for owner name")
+            }
+        }
+    }
+}
+
+/// Compute an [RFC 5155] NSEC3 hash using default settings.
+///
+/// See: [Nsec3param::default].
+///
+/// [RFC 5155]: https://www.rfc-editor.org/rfc/rfc5155
+pub fn nsec3_default_hash<N, HashOcts>(
+    owner: N,
+) -> Result<OwnerHash<HashOcts>, Nsec3HashError>
+where
+    N: ToName,
+    HashOcts: AsRef<[u8]> + EmptyBuilder + OctetsBuilder + Truncate,
+    for<'a> HashOcts: From<&'a [u8]>,
+{
+    let params = Nsec3param::<HashOcts>::default();
+    nsec3_hash(
+        owner,
+        params.hash_algorithm(),
+        params.iterations(),
+        params.salt(),
+    )
+}
+
+/// Compute an [RFC 5155] NSEC3 hash.
+///
+/// Computes an NSEC3 hash according to [RFC 5155] section 5:
+///
+/// > IH(salt, x, 0) = H(x || salt)
+/// > IH(salt, x, k) = H(IH(salt, x, k-1) || salt), if k > 0
+///
+/// Then the calculated hash of an owner name is:
+///
+/// > IH(salt, owner name, iterations),
+///
+/// Note that the `iterations` parameter is the number of _additional_
+/// iterations as defined in [RFC 5155] section 3.1.3.
+///
+/// [RFC 5155]: https://www.rfc-editor.org/rfc/rfc5155
+pub fn nsec3_hash<N, SaltOcts, HashOcts>(
+    owner: N,
+    algorithm: Nsec3HashAlg,
+    iterations: u16,
+    salt: &Nsec3Salt<SaltOcts>,
+) -> Result<OwnerHash<HashOcts>, Nsec3HashError>
+where
+    N: ToName,
+    SaltOcts: AsRef<[u8]>,
+    HashOcts: AsRef<[u8]> + EmptyBuilder + OctetsBuilder + Truncate,
+    for<'a> HashOcts: From<&'a [u8]>,
+{
+    if algorithm != Nsec3HashAlg::SHA1 {
+        return Err(Nsec3HashError::UnsupportedAlgorithm);
+    }
+
+    fn mk_hash<N, SaltOcts, HashOcts>(
+        owner: N,
+        iterations: u16,
+        salt: &Nsec3Salt<SaltOcts>,
+    ) -> Result<HashOcts, HashOcts::AppendError>
+    where
+        N: ToName,
+        SaltOcts: AsRef<[u8]>,
+        HashOcts: AsRef<[u8]> + EmptyBuilder + OctetsBuilder + Truncate,
+        for<'a> HashOcts: From<&'a [u8]>,
+    {
+        let mut canonical_owner = HashOcts::empty();
+        owner.compose_canonical(&mut canonical_owner)?;
+
+        let mut ctx = ring::digest::Context::new(&SHA1_FOR_LEGACY_USE_ONLY);
+        ctx.update(canonical_owner.as_ref());
+        ctx.update(salt.as_slice());
+        let mut h = ctx.finish();
+
+        for _ in 0..iterations {
+            let mut ctx =
+                ring::digest::Context::new(&SHA1_FOR_LEGACY_USE_ONLY);
+            ctx.update(h.as_ref());
+            ctx.update(salt.as_slice());
+            h = ctx.finish();
+        }
+
+        Ok(h.as_ref().into())
+    }
+
+    let hash = mk_hash(owner, iterations, salt)
+        .map_err(|_| Nsec3HashError::AppendError)?;
+
+    let owner_hash = OwnerHash::from_octets(hash)
+        .map_err(|_| Nsec3HashError::OwnerHashError)?;
+
+    Ok(owner_hash)
+}
+//============ Test ==========================================================
+
+#[cfg(test)]
+#[cfg(feature = "std")]
+mod test {
+    use super::*;
+    use crate::base::iana::{Class, Rtype};
+    use crate::base::Ttl;
+    use crate::rdata::dnssec::Timestamp;
+    use crate::rdata::{Mx, ZoneRecordData};
+    use crate::utils::base64;
+    use bytes::Bytes;
+    use std::str::FromStr;
+
+    type Name = crate::base::name::Name<Vec<u8>>;
+    type Ds = crate::rdata::Ds<Vec<u8>>;
+    type Dnskey = crate::rdata::Dnskey<Vec<u8>>;
+    type Rrsig = crate::rdata::Rrsig<Vec<u8>, Name>;
+
+    // Returns current root KSK/ZSK for testing (2048b)
+    fn root_pubkey() -> (Dnskey, Dnskey) {
+        let ksk = base64::decode::<Vec<u8>>(
+            "\
+            AwEAAaz/tAm8yTn4Mfeh5eyI96WSVexTBAvkMgJzkKTOiW1vkIbzxeF3+/\
+            4RgWOq7HrxRixHlFlExOLAJr5emLvN7SWXgnLh4+B5xQlNVz8Og8kvArMt\
+            NROxVQuCaSnIDdD5LKyWbRd2n9WGe2R8PzgCmr3EgVLrjyBxWezF0jLHwV\
+            N8efS3rCj/EWgvIWgb9tarpVUDK/b58Da+sqqls3eNbuv7pr+eoZG+SrDK\
+            6nWeL3c6H5Apxz7LjVc1uTIdsIXxuOLYA4/ilBmSVIzuDWfdRUfhHdY6+c\
+            n8HFRm+2hM8AnXGXws9555KrUB5qihylGa8subX2Nn6UwNR1AkUTV74bU=",
+        )
+        .unwrap();
+        let zsk = base64::decode::<Vec<u8>>(
+            "\
+            AwEAAeVDC34GZILwsQJy97K2Fst4P3XYZrXLyrkausYzSqEjSUulgh+iLgH\
+            g0y7FIF890+sIjXsk7KLJUmCOWfYWPorNKEOKLk5Zx/4M6D3IHZE3O3m/Ea\
+            hrc28qQzmTLxiMZAW65MvR2UO3LxVtYOPBEBiDgAQD47x2JLsJYtavCzNL5\
+            WiUk59OgvHmDqmcC7VXYBhK8V8Tic089XJgExGeplKWUt9yyc31ra1swJX5\
+            1XsOaQz17+vyLVH8AZP26KvKFiZeoRbaq6vl+hc8HQnI2ug5rA2zoz3MsSQ\
+            BvP1f/HvqsWxLqwXXKyDD1QM639U+XzVB8CYigyscRP22QCnwKIU=",
+        )
+        .unwrap();
+        (
+            Dnskey::new(257, 3, SecAlg::RSASHA256, ksk).unwrap(),
+            Dnskey::new(256, 3, SecAlg::RSASHA256, zsk).unwrap(),
+        )
+    }
+
+    // Returns the current net KSK/ZSK for testing (1024b)
+    fn net_pubkey() -> (Dnskey, Dnskey) {
+        let ksk = base64::decode::<Vec<u8>>(
+            "AQOYBnzqWXIEj6mlgXg4LWC0HP2n8eK8XqgHlmJ/69iuIHsa1TrHDG6TcOra/pyeGKwH0nKZhTmXSuUFGh9BCNiwVDuyyb6OBGy2Nte9Kr8NwWg4q+zhSoOf4D+gC9dEzg0yFdwT0DKEvmNPt0K4jbQDS4Yimb+uPKuF6yieWWrPYYCrv8C9KC8JMze2uT6NuWBfsl2fDUoV4l65qMww06D7n+p7RbdwWkAZ0fA63mXVXBZF6kpDtsYD7SUB9jhhfLQE/r85bvg3FaSs5Wi2BaqN06SzGWI1DHu7axthIOeHwg00zxlhTpoYCH0ldoQz+S65zWYi/fRJiyLSBb6JZOvn",
+        )
+        .unwrap();
+        let zsk = base64::decode::<Vec<u8>>(
+            "AQPW36Zs2vsDFGgdXBlg8RXSr1pSJ12NK+u9YcWfOr85we2z5A04SKQlIfyTK37dItGFcldtF7oYwPg11T3R33viKV6PyASvnuRl8QKiLk5FfGUDt1sQJv3S/9wT22Le1vnoE/6XFRyeb8kmJgz0oQB1VAO9b0l6Vm8KAVeOGJ+Qsjaq0O0aVzwPvmPtYm/i3qoAhkaMBUpg6RrF5NKhRyG3",
+        )
+        .unwrap();
+        (
+            Dnskey::new(257, 3, SecAlg::RSASHA256, ksk).unwrap(),
+            Dnskey::new(256, 3, SecAlg::RSASHA256, zsk).unwrap(),
+        )
+    }
+
+    #[test]
+    fn dnskey_digest() {
+        let (dnskey, _) = root_pubkey();
+        let owner = Name::root();
+        let expected = Ds::new(
+            20326,
+            SecAlg::RSASHA256,
+            DigestAlg::SHA256,
+            base64::decode::<Vec<u8>>(
+                "4G1EuAuPHTmpXAsNfGXQhFjogECbvGg0VxBCN8f47I0=",
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        assert_eq!(
+            dnskey.digest(&owner, DigestAlg::SHA256).unwrap().as_ref(),
+            expected.digest()
+        );
+    }
+
+    #[test]
+    fn dnskey_digest_unsupported() {
+        let (dnskey, _) = root_pubkey();
+        let owner = Name::root();
+        assert!(dnskey.digest(&owner, DigestAlg::GOST).is_err());
+    }
+
+    fn rrsig_verify_dnskey(ksk: Dnskey, zsk: Dnskey, rrsig: Rrsig) {
+        let mut records: Vec<_> = [&ksk, &zsk]
+            .iter()
+            .cloned()
+            .map(|x| {
+                Record::new(
+                    rrsig.signer_name().clone(),
+                    Class::IN,
+                    Ttl::from_secs(0),
+                    x.clone(),
+                )
+            })
+            .collect();
+        let signed_data = {
+            let mut buf = Vec::new();
+            rrsig.signed_data(&mut buf, records.as_mut_slice()).unwrap();
+            Bytes::from(buf)
+        };
+
+        // Test that the KSK is sorted after ZSK key
+        assert_eq!(ksk.key_tag(), rrsig.key_tag());
+        assert_eq!(ksk.key_tag(), records[1].data().key_tag());
+
+        // Test verifier
+        assert!(rrsig.verify_signed_data(&ksk, &signed_data).is_ok());
+        assert!(rrsig.verify_signed_data(&zsk, &signed_data).is_err());
+    }
+
+    #[test]
+    fn rrsig_verify_rsa_sha256() {
+        // Test 2048b long key
+        let (ksk, zsk) = root_pubkey();
+        let rrsig = Rrsig::new(
+            Rtype::DNSKEY,
+            SecAlg::RSASHA256,
+            0,
+            Ttl::from_secs(172800),
+            1560211200.into(),
+            1558396800.into(),
+            20326,
+            Name::root(),
+            base64::decode::<Vec<u8>>(
+                "otBkINZAQu7AvPKjr/xWIEE7+SoZtKgF8bzVynX6bfJMJuPay8jPvNmwXkZOdSoYlvFp0bk9JWJKCh8y5uoNfMFkN6OSrDkr3t0E+c8c0Mnmwkk5CETH3Gqxthi0yyRX5T4VlHU06/Ks4zI+XAgl3FBpOc554ivdzez8YCjAIGx7XgzzooEb7heMSlLc7S7/HNjw51TPRs4RxrAVcezieKCzPPpeWBhjE6R3oiSwrl0SBD4/yplrDlr7UHs/Atcm3MSgemdyr2sOoOUkVQCVpcj3SQQezoD2tCM7861CXEQdg5fjeHDtz285xHt5HJpA5cOcctRo4ihybfow/+V7AQ==",
+            )
+            .unwrap()
+        ).unwrap();
+        rrsig_verify_dnskey(ksk, zsk, rrsig);
+
+        // Test 1024b long key
+        let (ksk, zsk) = net_pubkey();
+        let rrsig = Rrsig::new(
+            Rtype::DNSKEY,
+            SecAlg::RSASHA256,
+            1,
+            Ttl::from_secs(86400),
+            Timestamp::from_str("20210921162830").unwrap(),
+            Timestamp::from_str("20210906162330").unwrap(),
+            35886,
+            "net.".parse::<Name>().unwrap(),
+            base64::decode::<Vec<u8>>(
+                "j1s1IPMoZd0mbmelNVvcbYNe2tFCdLsLpNCnQ8xW6d91ujwPZ2yDlc3lU3hb+Jq3sPoj+5lVgB7fZzXQUQTPFWLF7zvW49da8pWuqzxFtg6EjXRBIWH5rpEhOcr+y3QolJcPOTx+/utCqt2tBKUUy3LfM6WgvopdSGaryWdwFJPW7qKHjyyLYxIGx5AEuLfzsA5XZf8CmpUheSRH99GRZoIB+sQzHuelWGMQ5A42DPvOVZFmTpIwiT2QaIpid4nJ7jNfahfwFrCoS+hvqjK9vktc5/6E/Mt7DwCQDaPt5cqDfYltUitQy+YA5YP5sOhINChYadZe+2N80OA+RKz0mA==",
+            )
+            .unwrap()
+        ).unwrap();
+        rrsig_verify_dnskey(ksk, zsk, rrsig.clone());
+
+        // Test that 512b short RSA DNSKEY is not supported (too short)
+        let data = base64::decode::<Vec<u8>>(
+            "AwEAAcFcGsaxxdgiuuGmCkVImy4h99CqT7jwY3pexPGcnUFtR2Fh36BponcwtkZ4cAgtvd4Qs8PkxUdp6p/DlUmObdk=",
+        )
+        .unwrap();
+
+        let short_key = Dnskey::new(256, 3, SecAlg::RSASHA256, data).unwrap();
+        let err = rrsig
+            .verify_signed_data(&short_key, &vec![0; 100])
+            .unwrap_err();
+        assert_eq!(err, AlgorithmError::Unsupported);
+    }
+
+    #[test]
+    fn rrsig_verify_ecdsap256_sha256() {
+        let (ksk, zsk) = (
+            Dnskey::new(
+                257,
+                3,
+                SecAlg::ECDSAP256SHA256,
+                base64::decode::<Vec<u8>>(
+                    "mdsswUyr3DPW132mOi8V9xESWE8jTo0dxCjjnopKl+GqJxpVXckHAe\
+                    F+KkxLbxILfDLUT0rAK9iUzy1L53eKGQ==",
+                )
+                .unwrap(),
+            )
+            .unwrap(),
+            Dnskey::new(
+                256,
+                3,
+                SecAlg::ECDSAP256SHA256,
+                base64::decode::<Vec<u8>>(
+                    "oJMRESz5E4gYzS/q6XDrvU1qMPYIjCWzJaOau8XNEZeqCYKD5ar0IR\
+                    d8KqXXFJkqmVfRvMGPmM1x8fGAa2XhSA==",
+                )
+                .unwrap(),
+            )
+            .unwrap(),
+        );
+
+        let owner = Name::from_str("cloudflare.com.").unwrap();
+        let rrsig = Rrsig::new(
+            Rtype::DNSKEY,
+            SecAlg::ECDSAP256SHA256,
+            2,
+            Ttl::from_secs(3600),
+            1560314494.into(),
+            1555130494.into(),
+            2371,
+            owner,
+            base64::decode::<Vec<u8>>(
+                "8jnAGhG7O52wmL065je10XQztRX1vK8P8KBSyo71Z6h5wAT9+GFxKBaE\
+                zcJBLvRmofYFDAhju21p1uTfLaYHrg==",
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        rrsig_verify_dnskey(ksk, zsk, rrsig);
+    }
+
+    #[test]
+    fn rrsig_verify_ed25519() {
+        let (ksk, zsk) = (
+            Dnskey::new(
+                257,
+                3,
+                SecAlg::ED25519,
+                base64::decode::<Vec<u8>>(
+                    "m1NELLVVQKl4fHVn/KKdeNO0PrYKGT3IGbYseT8XcKo=",
+                )
+                .unwrap(),
+            )
+            .unwrap(),
+            Dnskey::new(
+                256,
+                3,
+                SecAlg::ED25519,
+                base64::decode::<Vec<u8>>(
+                    "2tstZAjgmlDTePn0NVXrAHBJmg84LoaFVxzLl1anjGI=",
+                )
+                .unwrap(),
+            )
+            .unwrap(),
+        );
+
+        let owner =
+            Name::from_octets(Vec::from(b"\x07ED25519\x02nl\x00".as_ref()))
+                .unwrap();
+        let rrsig = Rrsig::new(
+            Rtype::DNSKEY,
+            SecAlg::ED25519,
+            2,
+            Ttl::from_secs(3600),
+            1559174400.into(),
+            1557360000.into(),
+            45515,
+            owner,
+            base64::decode::<Vec<u8>>(
+                "hvPSS3E9Mx7lMARqtv6IGiw0NE0uz0mZewndJCHTkhwSYqlasUq7KfO5\
+                QdtgPXja7YkTaqzrYUbYk01J8ICsAA==",
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        rrsig_verify_dnskey(ksk, zsk, rrsig);
+    }
+
+    #[test]
+    fn rrsig_verify_generic_type() {
+        let (ksk, zsk) = root_pubkey();
+        let rrsig = Rrsig::new(
+            Rtype::DNSKEY,
+            SecAlg::RSASHA256,
+            0,
+            Ttl::from_secs(172800),
+            1560211200.into(),
+            1558396800.into(),
+            20326,
+            Name::root(),
+            base64::decode::<Vec<u8>>(
+                "otBkINZAQu7AvPKjr/xWIEE7+SoZtKgF8bzVynX6bfJMJuPay8jPvNmwXkZ\
+                OdSoYlvFp0bk9JWJKCh8y5uoNfMFkN6OSrDkr3t0E+c8c0Mnmwkk5CETH3Gq\
+                xthi0yyRX5T4VlHU06/Ks4zI+XAgl3FBpOc554ivdzez8YCjAIGx7XgzzooE\
+                b7heMSlLc7S7/HNjw51TPRs4RxrAVcezieKCzPPpeWBhjE6R3oiSwrl0SBD4\
+                /yplrDlr7UHs/Atcm3MSgemdyr2sOoOUkVQCVpcj3SQQezoD2tCM7861CXEQ\
+                dg5fjeHDtz285xHt5HJpA5cOcctRo4ihybfow/+V7AQ==",
+            )
+            .unwrap(),
+        )
+        .unwrap();
+
+        let mut records: Vec<Record<Name, ZoneRecordData<Vec<u8>, Name>>> =
+            [&ksk, &zsk]
+                .iter()
+                .cloned()
+                .map(|x| {
+                    let data = ZoneRecordData::from(x.clone());
+                    Record::new(
+                        rrsig.signer_name().clone(),
+                        Class::IN,
+                        Ttl::from_secs(0),
+                        data,
+                    )
+                })
+                .collect();
+
+        let signed_data = {
+            let mut buf = Vec::new();
+            rrsig.signed_data(&mut buf, records.as_mut_slice()).unwrap();
+            Bytes::from(buf)
+        };
+
+        assert!(rrsig.verify_signed_data(&ksk, &signed_data).is_ok());
+    }
+
+    #[test]
+    fn rrsig_verify_wildcard() {
+        let key = Dnskey::new(
+            256,
+            3,
+            SecAlg::RSASHA1,
+            base64::decode::<Vec<u8>>(
+                "AQOy1bZVvpPqhg4j7EJoM9rI3ZmyEx2OzDBVrZy/lvI5CQePxX\
+                HZS4i8dANH4DX3tbHol61ek8EFMcsGXxKciJFHyhl94C+NwILQd\
+                zsUlSFovBZsyl/NX6yEbtw/xN9ZNcrbYvgjjZ/UVPZIySFNsgEY\
+                vh0z2542lzMKR4Dh8uZffQ==",
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        let rrsig = Rrsig::new(
+            Rtype::MX,
+            SecAlg::RSASHA1,
+            2,
+            Ttl::from_secs(3600),
+            Timestamp::from_str("20040509183619").unwrap(),
+            Timestamp::from_str("20040409183619").unwrap(),
+            38519,
+            Name::from_str("example.").unwrap(),
+            base64::decode::<Vec<u8>>(
+                "OMK8rAZlepfzLWW75Dxd63jy2wswESzxDKG2f9AMN1CytCd10cYI\
+                 SAxfAdvXSZ7xujKAtPbctvOQ2ofO7AZJ+d01EeeQTVBPq4/6KCWhq\
+                 e2XTjnkVLNvvhnc0u28aoSsG0+4InvkkOHknKxw4kX18MMR34i8lC\
+                 36SR5xBni8vHI=",
+            )
+            .unwrap(),
+        )
+        .unwrap();
+        let record = Record::new(
+            Name::from_str("a.z.w.example.").unwrap(),
+            Class::IN,
+            Ttl::from_secs(3600),
+            Mx::new(1, Name::from_str("ai.example.").unwrap()),
+        );
+        let signed_data = {
+            let mut buf = Vec::new();
+            rrsig.signed_data(&mut buf, &mut [record]).unwrap();
+            Bytes::from(buf)
+        };
+
+        // Test that the key matches RRSIG
+        assert_eq!(key.key_tag(), rrsig.key_tag());
+
+        // Test verifier
+        assert_eq!(rrsig.verify_signed_data(&key, &signed_data), Ok(()));
+    }
+}
