+++ src/dnssec/sign/config.rs
@@ -3,16 +3,14 @@ use core::marker::PhantomData;
 
 use octseq::{EmptyBuilder, FromBuilder};
 
+use super::denial::config::DenialConfig;
+use super::denial::nsec3::{Nsec3HashProvider, OnDemandNsec3HashProvider};
+use super::records::{DefaultSorter, Sorter};
+use super::signatures::strategy::DefaultSigningKeyUsageStrategy;
+use super::signatures::strategy::RrsigValidityPeriodStrategy;
+use super::signatures::strategy::SigningKeyUsageStrategy;
 use crate::base::{Name, ToName};
-use crate::sign::denial::config::DenialConfig;
-use crate::sign::denial::nsec3::{
-    Nsec3HashProvider, OnDemandNsec3HashProvider,
-};
-use crate::sign::records::{DefaultSorter, Sorter};
-use crate::sign::signatures::strategy::DefaultSigningKeyUsageStrategy;
-use crate::sign::signatures::strategy::RrsigValidityPeriodStrategy;
-use crate::sign::signatures::strategy::SigningKeyUsageStrategy;
-use crate::sign::SignRaw;
+use crate::crypto::misc::SignRaw;
 
 //------------ SigningConfig -------------------------------------------------
