+++ src/dnssec/sign/keys/keyset.rs
@@ -7,7 +7,7 @@
 //!
 //! ```no_run
 //! use domain::base::Name;
-//! use domain::sign::keys::keyset::{KeySet, RollType, UnixTime};
+//! use domain::dnssec::sign::keys::keyset::{KeySet, RollType, UnixTime};
 //! use std::fs::File;
 //! use std::io::Write;
 //! use std::str::FromStr;
@@ -1517,7 +1517,7 @@ fn csk_roll_actions(rollstate: RollState) -> Vec<Action> {
 #[cfg(test)]
 mod tests {
     use crate::base::Name;
-    use crate::sign::keys::keyset::{
+    use crate::dnssec::sign::keys::keyset::{
         Action, KeySet, KeyType, RollType, UnixTime,
     };
     use crate::std::string::ToString;
