+++ src/dnssec/validator/group.rs
@@ -17,12 +17,12 @@ use crate::base::name::ToName;
 use crate::base::opt::exterr::ExtendedError;
 use crate::base::rdata::ComposeRecordData;
 use crate::base::{Name, ParsedName, ParsedRecord, Record, Rtype, Ttl};
+use crate::crypto::validate::RrsigExt;
 use crate::dep::octseq::builder::with_infallible;
 use crate::dep::octseq::{Octets, OctetsFrom};
 use crate::net::client::request::{RequestMessage, SendRequest};
 use crate::rdata::dnssec::Timestamp;
 use crate::rdata::{AllRecordData, Dnskey, Rrsig};
-use crate::validate::RrsigExt;
 use bytes::Bytes;
 use moka::future::Cache;
 use ring::digest;
